class name refactoring
